+++ src/assembler/linker.rs
@@ -12,13 +12,10 @@ pub enum Error {
     LocalBeforeGlobal(String),
 }
 
-pub type Globals = HashMap<String, u16>;
-pub type Locals = HashMap<String, u16>;
-
-pub fn link(ast: &[ParsedItem]) -> Result<Vec<u16>, Error> {
+pub fn link(ast: &[ParsedItem]) -> Result<(Vec<u16>, Globals), Error> {
 
     let mut bin = Vec::new();
-    let (mut globals, mut locals) = try!(extract_labels(ast));
+    let mut labels = try!(extract_labels(ast));
     let mut changed = true;
 
     while changed {
@@ -29,43 +26,39 @@ pub fn link(ast: &[ParsedItem]) -> Result<Vec<u16>, Error> {
         for item in ast {
             match *item {
                 ParsedItem::Directive(Directive::Lcomm(ref symbol, size)) => {
-                    let ptr = globals.get_mut(symbol).unwrap();
-                    if *ptr != index {
-                        *ptr = index;
+                    let label = labels.get_mut(symbol).unwrap();
+                    if label.addr != index {
+                        label.addr = index;
                         changed = true;
                     }
-                    last_global = Some(symbol);
+                    last_global = Some(symbol.clone());
 
                     bin.extend(iter::repeat(0).take(size as usize));
                     index += size;
                 }
-                ParsedItem::Directive(ref d) => index += match last_global {
-                    Some(s) => try!(d.append_to(&mut bin, &globals, &locals.get(s).unwrap())),
-                    None => try!(d.append_to(&mut bin, &globals, &HashMap::new())),
-                },
+                ParsedItem::Directive(ref d) =>
+                    index += try!(d.append_to(&mut bin, &labels, &last_global)),
                 ParsedItem::LabelDecl(ref s) => {
-                    let ptr = globals.get_mut(s).unwrap();
-                    if *ptr != index {
-                        *ptr = index;
+                    let label = labels.get_mut(s).unwrap();
+                    if label.addr != index {
+                        label.addr = index;
                         changed = true;
                     }
-                    last_global = Some(s);
+                    last_global = Some(s.clone());
                 }
                 ParsedItem::LocalLabelDecl(ref s) => {
-                    let ptr = locals.get_mut(*last_global.as_ref().unwrap())
-                                    .unwrap()
-                                    .get_mut(s)
-                                    .unwrap();
-                    if *ptr != index {
+                    let addr = labels.get_mut(last_global.as_ref().unwrap())
+                                     .unwrap()
+                                     .locals
+                                     .get_mut(s)
+                                     .unwrap();
+                    if *addr != index {
                         changed = true;
-                        *ptr = index;
+                        *addr = index;
                     }
                 }
                 ParsedItem::Instruction(ref i) => {
-                    let solved = match last_global {
-                        Some(s) => try!(i.solve(&globals, locals.get(s).unwrap())),
-                        None => try!(i.solve(&globals, &HashMap::new())),
-                    };
+                    let solved = try!(i.solve(&labels, &last_global));
                     bin.extend_from_slice(&[0xbeaf; 3]);
                     index += solved.encode(&mut bin[index as usize..]);
                     bin.truncate(index as usize);
@@ -75,32 +68,30 @@ pub fn link(ast: &[ParsedItem]) -> Result<Vec<u16>, Error> {
         }
     }
 
-    Ok(bin)
+    Ok((bin, labels))
 }
 
-fn extract_labels
-    (ast: &[ParsedItem])
-     -> Result<(Globals, HashMap<String, Locals>), Error> {
+fn extract_labels(ast: &[ParsedItem]) -> Result<Globals, Error> {
     let mut prev_label = None;
-    let mut globals = HashMap::new();
-    let mut locals = HashMap::new();
+    let mut labels = HashMap::new();
 
     for item in ast.iter() {
         match *item {
             ParsedItem::LabelDecl(ref s) | ParsedItem::Directive(Directive::Lcomm(ref s, _)) => {
                 prev_label = Some(s.clone());
-                if globals.contains_key(s) {
+                if labels.contains_key(s) {
                     return Err(Error::DuplicatedLabel(s.clone()));
                 } else {
-                    globals.insert(s.clone(), 0);
-                    locals.insert(s.clone(), HashMap::new());
+                    labels.insert(s.clone(), LabelInfos::default());
                 }
             }
             ParsedItem::LocalLabelDecl(ref s) => {
                 if prev_label.is_none() {
                     return Err(Error::LocalBeforeGlobal(s.clone()));
                 }
-                let locals = locals.get_mut(prev_label.as_ref().unwrap()).unwrap();
+                let locals = &mut labels.get_mut(prev_label.as_ref().unwrap())
+                                        .unwrap()
+                                        .locals;
                 if locals.contains_key(s) {
                     return Err(Error::DuplicatedLocalLabel(s.clone()));
                 } else {
@@ -111,5 +102,5 @@ fn extract_labels
         }
     }
 
-    Ok((globals, locals))
+    Ok(labels)
 }